+++ components/datadog/agent/package.go
@@ -1,4 +1,4 @@
-package utils
+package agent
 
 import (
 	"fmt"
